Rewrite mingw code to use the wincrypt API
Even though that API is flagged as deprecated, it is still in
use and is certainly fully available before Vista.

Since it seems bcrypt.h should be okay in the way in which
we used it _after_ Vista, it means we can upgrade if/when we
decide to ditch support for Vista.

Since most MK code around assumes Vista, I have decided to
make this piece of code not an exception to that.

Prodded by @darkk's comments (thanks!).